update compilation for error flag

+++ srcs/redirections/check_redirections_part1.c
@@ -2,7 +2,7 @@
 
 int		is_redirection(t_env *e, int i)
 {
-	if (i >= e->len_mag)
+	if (i >= (int)e->len_mag)
 		return (0);
 	if (e->magic[i].type && !ft_strcmp(e->magic[i].type, "red"))
 		return (1);
@@ -11,7 +11,7 @@ int		is_redirection(t_env *e, int i)
 
 int		is_output_redir(t_env *e, int i)
 {
-	if (i >= e->len_mag)
+	if (i >= (int)e->len_mag)
 		return (0);
 	if (!ft_strcmp(e->magic[i].cmd, ">") ||
 		!ft_strcmp(e->magic[i].cmd, ">>"))
@@ -21,7 +21,7 @@ int		is_output_redir(t_env *e, int i)
 
 int		is_input_redir(t_env *e, int i)
 {
-	if (i >= e->len_mag)
+	if (i >= (int)e->len_mag)
 		return (0);
 	if (!ft_strcmp(e->magic[i].cmd, "<"))
 		return (1);
@@ -30,7 +30,7 @@ int		is_input_redir(t_env *e, int i)
 
 int		is_heredoc(t_env *e, int i)
 {
-	if (i >= e->len_mag)
+	if (i >= (int)e->len_mag)
 		return (0);
 	if (!ft_strcmp(e->magic[i].cmd, "<<"))
 		return (1);
@@ -39,7 +39,7 @@ int		is_heredoc(t_env *e, int i)
 
 int		is_aggregator(t_env *e, int i)
 {
-	if (i >= e->len_mag)
+	if (i >= (int)e->len_mag)
 		return (0);
 	if (ft_strstr(e->magic[i].cmd, "<&") ||
 		ft_strstr(e->magic[i].cmd, ">&"))
